% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCIFitnessCoverageScore.R
\name{fCIFitnessCoverageScore}
\alias{fCIFitnessCoverageScore}
\title{Fitness function to maximise the coverage}
\usage{
fCIFitnessCoverageScore(
  viSelected,
  nRadiusUpperLimit_m = 6000,
  iMaxStations = Inf,
  dtLocations = NULL,
  dtWays = NULL,
  dtPolygonCoveredEdges = NULL
)
}
\arguments{
\item{viSelected}{A vector of 0s and 1s to mark out which of the
status = ? stations are selected in this solution}

\item{nRadiusUpperLimit_m}{Radius to be consiered to calculate coverage}

\item{iMaxStations}{The max stations constraints imposed above which soln is invalid}
}
\description{
Fitness function to maximise the coverage
}
