% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCIFitnessStations.R
\name{fCIFitnessStations}
\alias{fCIFitnessStations}
\title{Fitness function to minimise the number of stations}
\usage{
fCIFitnessStations(
  viSelected,
  nRadiusUpperLimit_m = 6000,
  nMinCoverage_pct = 100,
  dtLocations = NULL,
  dtWays = NULL,
  dtPolygonCoveredEdges = NULL
)
}
\arguments{
\item{viSelected}{A vector of 0s and 1s to mark out which of the
status = ? stations are selected in this solution}

\item{nRadiusUpperLimit_m}{Radius to be consiered to calculate coverage}

\item{nMinCoverage_pct}{Coverage min constraint imposed below which soln is invalid}
}
\description{
Fitness function to minimise the number of stations
}
