% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCustomMutation.R
\name{fCustomMutation}
\alias{fCustomMutation}
\title{Mutating the organisms}
\usage{
fCustomMutation(object, parent, dtLocations, dtWays, dtPolygonCoveredEdges)
}
\arguments{
\item{object:}{The GA object thing}

\item{parent:}{GA also insists on another argument, which is the specific organism}
}
\description{
Mutating the organisms
}
