% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCustomPMutation.R
\name{fCustomPMutation}
\alias{fCustomPMutation}
\title{Mutation probability of the GA organisms}
\usage{
fCustomPMutation(object, iUpdateAfterEvery = 0, i = 0, iMaxIterations)
}
\description{
Mutation probability of the GA organisms
}
