% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fEdgeSpan.R
\name{fEdgeSpan}
\alias{fEdgeSpan}
\title{Function to calculate what edges fall under what node's radius}
\usage{
fEdgeSpan(
  igraphLinks,
  cDestinationNodeID,
  dtNodes,
  dtWays,
  nEarthCircumference_m = 4e+07,
  nDistanceCutoff_m = 6000
)
}
\arguments{
\item{igraphLinks}{Graph from dtWays. Look through server.r to find creation}

\item{cDestinationNodeID}{The node ID around which we need to find the coverage for.
Note that that is the destination and not the source.}
}
\description{
Function to calculate what edges fall under what node's radius
}
