% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fEnclosingNodePolygon.R
\name{fEnclosingNodePolygon}
\alias{fEnclosingNodePolygon}
\title{Find the enclosing node polygon. That's what the edge span is calculated on.}
\usage{
fEnclosingNodePolygon(
  vcPointsID,
  vnPointsLatitude_deg,
  vnPointsLongitude_deg,
  vNodeID,
  vnNodeLongitude_deg,
  vnNodeLatitude_deg,
  nMaxVertexDistanceFromPoint_m = 250
)
}
\arguments{
\item{vcPointsID}{Location IDs of the points}

\item{vnPointsLatitude_deg}{Respective Lats of the location IDs}

\item{vnPointsLongitude_deg}{Respective Longs of the location IDs}

\item{vNodeID}{Node IDs of the points}

\item{vnNodeLongitude_deg}{Respective Lats of the Node IDs}

\item{vnNodeLatitude_deg}{Respective Longs of the Node IDs}

\item{nMaxVertexDistanceFromPoint_m}{If the polygon nodes are going beyond
this then take the closest point instead.}
}
\description{
Find the enclosing node polygon. That's what the edge span is calculated on.
}
