% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGenerateCoverageSummary.R
\name{fGenerateCoverageSummary}
\alias{fGenerateCoverageSummary}
\title{Get the road length covered within various bucketed radii}
\usage{
fGenerateCoverageSummary(
  dtLocations,
  dtWays,
  dtPolygonCoveredEdges,
  nMaxRadiusUserWillUse_m = nMaxRadiusUserWillUse_m
)
}
\description{
Get the road length covered within various bucketed radii
}
