% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetCityPlot.R
\name{fGetCityPlot}
\alias{fGetCityPlot}
\title{Makes a ggMap plot based on the selected scenario}
\usage{
fGetCityPlot(
  ggMapTiles = NULL,
  selectizeInputPlotScenario = NULL,
  dtOptimisedLocations = NULL,
  dtLocations = NULL,
  dtWays = NULL,
  dtOptimisationCoverageSummmary = NULL,
  dtCoveredEdges = NULL,
  sliderLocationSize = 1,
  sliderCoverageSize = 1
)
}
\description{
Makes a ggMap plot based on the selected scenario
}
