% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetDisconnectedPaths.R
\name{fGetDisconnectedPaths}
\alias{fGetDisconnectedPaths}
\title{dtRandomGrid is a set of random points from across the area
We check if each path in the data can reach points in dtRandomGrid
Paths which can't reach more than a certain number of points are excluded
The brute force way would be to check each path against each path but
that's too time consuming. We'll live with this heuristic.}
\usage{
fGetDisconnectedPaths(dtNodes, dtWays)
}
\description{
dtRandomGrid is a set of random points from across the area
We check if each path in the data can reach points in dtRandomGrid
Paths which can't reach more than a certain number of points are excluded
The brute force way would be to check each path against each path but
that's too time consuming. We'll live with this heuristic.
}
