% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fNearestLocation.R
\name{fNearestLocation}
\alias{fNearestLocation}
\title{A function to find the nearest location from a certain x, y
It returns an integer to pick the position in the vector of locations}
\usage{
fNearestLocation(
  nLongitude_deg = 0,
  nLatitude_deg = 0,
  nRadius_m = 1000,
  vnLongitude_deg = 0,
  vnLatitude_deg = 0
)
}
\arguments{
\item{nRadius_m}{radius within which to search around x, y}
}
\description{
A function to find the nearest location from a certain x, y
It returns an integer to pick the position in the vector of locations
}
